+++ rmdev.ibge.localidades/Contracts/IIBGELocalidades_UF.cs
@@ -13,7 +13,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoUF">Identificador da Unidade da Federação</param>
         /// <returns>Dados da Unidade da Federação</returns>
         [Get("/api/v1/localidades/estados/{codigoUF}")]
-        Task<UF> BuscarUFAsync(int codigoUF);
+        Task<UF> BuscarUFAsync(long codigoUF);
 
         /// <summary>
         /// Obtém o conjunto de Unidades da Federação do Brasil
@@ -21,7 +21,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoUFs">Identificadores de Unidades da Federação</param>
         /// <returns>Lista de Unidades da Federação</returns>
         [Get("/api/v1/localidades/estados/{codigoUFs}")]
-        Task<List<UF>> BuscarUFsAsync(params int[] codigoUFs);
+        Task<List<UF>> BuscarUFsAsync(params long[] codigoUFs);
 
         /// <summary>
         /// Obtém o conjunto de Unidades da Federação do Brasil
@@ -29,7 +29,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoMacrorregiao">Um ou mais identificadores de regiões</param>
         /// <returns>Lista de Unidades da Federação</returns>
         [Get("/api/v1/localidades/regioes/{codigoMacrorregiao}/estados")]
-        Task<List<UF>> BuscarUFsPorMacrorregiaoAsync(params int[] codigoMacrorregiao);
+        Task<List<UF>> BuscarUFsPorMacrorregiaoAsync(params long[] codigoMacrorregiao);
 
     }
 }
